Identity And morder modules added

+++ Shared/Shared/Domain/ValueObjects/Address.cs
@@ -1,5 +1,5 @@
 namespace Shared.Domain.ValueObjects;
 
-public record Address(string City, string Street, uint ZipCode, short HouseNumber, string? ExtraDetails)
+public record Address(string City, string Street, uint ZipCode, uint HouseNumber, string? ExtraDetails)
 {
 }
